requirements.txt
pdfminer.six
rich==13.7.0